Fix namespaces and rename directories to follow PSR4
