+++ tests/Integration/Endpoints/PaymentTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Alma\API\Tests\Integration;
+namespace Alma\API\Tests\Integration\Endpoints;
 
 use Alma\API\Client;
 use Alma\API\Entities\Instalment;
@@ -8,8 +8,7 @@ use Alma\API\Entities\Payment;
 use PHPUnit\Framework\TestCase;
 
 final class PaymentsTest extends TestCase
-{
-    protected static $almaClient;
+{    protected static $almaClient;
 
     public static function setUpBeforeClass()
     {
