+++ tests/Unit/ClientContextTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Alma\API\Tests;
+namespace Alma\API\Tests\Unit;
 
 use Alma\API\ClientContext;
 use PHPUnit\Framework\TestCase;
