+++ tests/Unit/ClientOptionsValidatorTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Alma\API\Tests;
+namespace Alma\API\Tests\Unit;
 
 use Alma\API\Lib\ClientOptionsValidator;
 use Alma\API\Client;
